pyproject.toml
[tool.poetry]
name = "phoenix_rising"
version = "0.1.0"
description = "A sanctuary against corporate darkness - rise like a phoenix"
authors = ["Ericson Willians <ericsonwillians@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^1.32.0"
huggingface-hub = "^0.20.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
aiosqlite = "^0.19.0"
requests = "^2.31.0"
pydantic = "^2.6.1"

[tool.poetry.dev-dependencies]
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"